Rename `cupy.logic` submodule to `cupy._logic`
