+++ cupy/_logic/content.py
