+++ cupy/_logic/truth.py
