+++ cupy/_logic/type_test.py
