--- cupy/logic/__init__.py
@@ -1,9 +1,0 @@
-# Functions from the following NumPy document
-# https://docs.scipy.org/doc/numpy/reference/routines.logic.html
-
-# "NOQA" to suppress flake8 warning
-from cupy.logic import comparison  # NOQA
-from cupy.logic import content  # NOQA
-from cupy.logic import ops  # NOQA
-from cupy.logic import truth  # NOQA
-from cupy.logic import type_test  # NOQA
